Update 02-novice.md
In the instructor training, I struggled with the phrase used in line 195 of this file - "the world is only a few thousand years old". The software carpentry ethos is very much centred on diversity but my feeling when this slide came up was that anybody with a very traditional Christian / Muslim / Judaism belief system (i.e. reading the old testament literally) would have felt belittled by this example. Especially with the humorous reaction of the class to such an idea. I would like to propose that replacing the line with "the world is flat" would achieve the same overall message but without risking offending anybody's views.